Update to take advantage of downstream changes.
Add clients, oauth2, and sessions support.

Add a middleware for embedding a logger in a request.

Add a middleware for embedding a token in a request, and adding its info
to the logger for that request.

Add module replacements for local versions of modules, like a sociopath.

Load our JWT PEM key and use that instead of a secret, now that tokens
and sessions are on RS256 instead of HS256.

Pull acceptable Google client IDs from an environment variable.